pyproject.toml
[tool.poetry]
name = "money"
version = "0.1.0"
description = ""
authors = ["PingWinCheck <ankolt@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
pydantic = {extras = ["email"], version = "^2.8.2"}
python-multipart = "^0.0.9"
uvicorn = "^0.30.5"
jinja2 = "^3.1.4"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
psycopg2 = "^2.9.9"
bcrypt = "^4.2.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
